requirements.txt
mediapipe
tqdm
torch
torchvision
tokenizers
evaluate
sacrebleu
numpy
pandas
opencv-python
pyyaml
